file upload and download is done

+++ Backend/project/src/main/java/idpa/project/repository/FileDataRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FileDataRepository extends JpaRepository<FileData, Integer> {
+public interface FileDataRepository extends JpaRepository<FileData, Long> {
 }
